refactor(fixture): split zome into integrity/coordinator
