+++ ts/test/fixture/zomes/coordinator/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
-name = "crud"
 edition = "2021"
+name = "coordinator"
 version = "0.1.0"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
-name = "crud"
+name = "coordinator"
 
 [dependencies]
 hdk = "0.0.140"
+integrity = { path = "../integrity" }
 serde = "1"
