+++ ts/test/fixture/zomes/integrity/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+edition = "2021"
+name = "integrity"
+version = "0.1.0"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+name = "integrity"
+
+[dependencies]
+holochain_deterministic_integrity = "0.0.12"
+serde = "1"
