+++ ts/test/fixture/zomes/integrity/src/lib.rs
@@ -1,0 +1,16 @@
+use holochain_deterministic_integrity::prelude::*;
+
+#[hdk_entry_helper]
+pub struct Content(pub String);
+
+#[hdk_entry_helper]
+pub struct UpdateInput {
+    pub hash: ActionHash,
+    pub content: String,
+}
+
+#[hdk_entry_defs]
+#[unit_enum(EntryTypesUnit)]
+pub enum EntryTypes {
+    Content(Content),
+}
