+++ ts/test/local/conductor.ts
@@ -187,14 +187,14 @@ test("Local Conductor - Install and call a hApp bundle", async (t) => {
   const entryContent = "Bye bye, world";
   const createEntryResponse: EntryHash =
     await installedHappBundle.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "create",
       payload: entryContent,
     });
   t.ok(createEntryResponse);
   const readEntryResponse: typeof entryContent =
     await installedHappBundle.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       payload: createEntryResponse,
     });
@@ -209,10 +209,14 @@ test("Local Conductor - Get a convenience function for zome calls", async (t) =>
   const [aliceHapps] = await conductor.installAgentsHapps({
     agentsDnas: [[{ path: FIXTURE_DNA_URL.pathname }]],
   });
-  const crudZomeCall = getZomeCaller(aliceHapps.cells[0], "crud");
-  t.equal(typeof crudZomeCall, "function", "getZomeCaller returns a function");
+  const coordinatorZomeCall = getZomeCaller(aliceHapps.cells[0], "coordinator");
+  t.equal(
+    typeof coordinatorZomeCall,
+    "function",
+    "getZomeCaller returns a function"
+  );
 
-  const entryHeaderHash: ActionHash = await crudZomeCall(
+  const entryHeaderHash: ActionHash = await coordinatorZomeCall(
     "create",
     "test-entry"
   );
@@ -262,7 +266,9 @@ test("Local Conductor - Zome call can time out before completion", async (t) => 
   const cell = aliceHapp.namedCells.get("test");
   assert(cell);
 
-  await t.rejects(cell.callZome({ fn_name: "create", zome_name: "crud" }, 1));
+  await t.rejects(
+    cell.callZome({ fn_name: "create", zome_name: "coordinator" }, 1)
+  );
 
   await conductor.shutDown();
   await cleanAllConductors();
@@ -298,7 +304,7 @@ test("Local Conductor - Create and read an entry using the entry zome", async (t
   const createEntryHash: EntryHash = await conductor.appWs().callZome({
     cap_secret: null,
     cell_id,
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     provenance: agentPubKey,
     payload: entryContent,
@@ -312,7 +318,7 @@ test("Local Conductor - Create and read an entry using the entry zome", async (t
     .callZome({
       cap_secret: null,
       cell_id,
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       provenance: agentPubKey,
       payload: createEntryHash,
@@ -336,7 +342,7 @@ test("Local Conductor - Create and read an entry using the entry zome, 2 conduct
 
   const entryContent = "test-content";
   const createEntryHash: EntryHash = await aliceHapps.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     payload: entryContent,
   });
@@ -348,7 +354,7 @@ test("Local Conductor - Create and read an entry using the entry zome, 2 conduct
 
   const readEntryResponse: typeof entryContent =
     await bobHapps.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       payload: createEntryHash,
     });
@@ -376,7 +382,7 @@ test("Local Conductor - Receive a signal", async (t) => {
 
   const aliceSignal = { value: "signal" };
   aliceHapps.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     payload: aliceSignal,
   });
