+++ ts/test/local/scenario.ts
@@ -71,7 +71,7 @@ test("Local Scenario - runScenario - Catch error that occurs in a signal handler
 
     const signalAlice = { value: "hello alice" };
     alice.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "signal_loopback",
       payload: signalAlice,
     });
@@ -114,7 +114,7 @@ test("Local Scenario - Create and read an entry, 2 conductors", async (t) => {
 
   const content = "Hi dare";
   const createEntryHash = await alice.cells[0].callZome<EntryHash>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     payload: content,
   });
@@ -122,7 +122,7 @@ test("Local Scenario - Create and read an entry, 2 conductors", async (t) => {
   await pause(100);
 
   const readContent = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
@@ -142,7 +142,7 @@ test("Local Scenario - Conductor maintains data after shutdown and restart", asy
 
   const content = "Before shutdown";
   const createEntryHash = await alice.cells[0].callZome<EntryHash>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     payload: content,
   });
@@ -150,7 +150,7 @@ test("Local Scenario - Conductor maintains data after shutdown and restart", asy
   await pause(100);
 
   const readContent = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
@@ -162,7 +162,7 @@ test("Local Scenario - Conductor maintains data after shutdown and restart", asy
   await bob.conductor.startUp();
   await bob.conductor.connectAppInterface();
   const readContentAfterRestart = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
@@ -196,13 +196,13 @@ test("Local Scenario - Receive signals with 2 conductors", async (t) => {
 
   const signalAlice = { value: "hello alice" };
   alice.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     payload: signalAlice,
   });
   const signalBob = { value: "hello bob" };
   bob.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     payload: signalBob,
   });
