+++ ts/test/trycp/conductor.ts
@@ -81,14 +81,14 @@ test("TryCP Conductor - Install and call a hApp bundle", async (t) => {
   const entryContent = "Bye bye, world";
   const createEntryResponse: EntryHash =
     await installedHappBundle.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "create",
       payload: entryContent,
     });
   t.ok(createEntryResponse);
   const readEntryResponse: typeof entryContent =
     await installedHappBundle.cells[0].callZome({
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       payload: createEntryResponse,
     });
@@ -127,7 +127,7 @@ test("TryCP Conductor - Receive a signal", async (t) => {
   conductor.appWs().callZome({
     cap_secret: null,
     cell_id: installedAppBundle.cell_data[0].cell_id,
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     provenance: agentPubKey,
     payload: testSignal,
@@ -179,7 +179,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome", async (t
   const createEntryHash = await conductor.appWs().callZome<EntryHash>({
     cap_secret: null,
     cell_id,
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     provenance: agentPubKey,
     payload: entryContent,
@@ -193,7 +193,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome", async (t
     .callZome<typeof entryContent>({
       cap_secret: null,
       cell_id,
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       provenance: agentPubKey,
       payload: createEntryHash,
@@ -222,7 +222,7 @@ test("TryCP Conductor - Reading a non-existent entry returns null", async (t) =>
   await conductor.connectAppInterface();
 
   const actual = await alice_happs.cells[0].callZome<null>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     provenance: alice_happs.agentPubKey,
     payload: Buffer.from("hCkk", "base64"),
@@ -301,7 +301,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome, 1 conduct
   const createEntryHash = await conductor.appWs().callZome<EntryHash>({
     cap_secret: null,
     cell_id: cellId1,
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     provenance: agent1PubKey,
     payload: entryContent,
@@ -315,7 +315,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome, 1 conduct
     .callZome<typeof entryContent>({
       cap_secret: null,
       cell_id: cellId2,
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       provenance: agent2PubKey,
       payload: createEntryHash,
@@ -349,7 +349,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome, 2 conduct
   const createEntry1Hash = await conductor1.appWs().callZome<EntryHash>({
     cap_secret: null,
     cell_id: alice.cells[0].cell_id,
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     provenance: alice.agentPubKey,
     payload: entryContent,
@@ -365,7 +365,7 @@ test("TryCP Conductor - Create and read an entry using the entry zome, 2 conduct
     .callZome<typeof entryContent>({
       cap_secret: null,
       cell_id: bob.cells[0].cell_id,
-      zome_name: "crud",
+      zome_name: "coordinator",
       fn_name: "read",
       provenance: bob.agentPubKey,
       payload: createEntry1Hash,
