+++ ts/test/trycp/scenario.ts
@@ -62,13 +62,13 @@ test("TryCP Scenario - Receive signals with 2 conductors", async (t) => {
 
   const signalAlice = { value: "hello alice" };
   alice.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     payload: signalAlice,
   });
   const signalBob = { value: "hello bob" };
   bob.cells[0].callZome({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "signal_loopback",
     payload: signalBob,
   });
@@ -95,7 +95,7 @@ test("TryCp Scenario - Create and read an entry, 2 conductors", async (t) => {
 
   const content = "Hi dare";
   const createEntryHash = await alice.cells[0].callZome<EntryHash>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     payload: content,
   });
@@ -103,7 +103,7 @@ test("TryCp Scenario - Create and read an entry, 2 conductors", async (t) => {
   await pause(100);
 
   const readContent = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
@@ -121,13 +121,13 @@ test("TryCP Scenario - Conductor maintains data after shutdown and restart", asy
   await scenario.shareAllAgents();
   const content = "Before shutdown";
   const createEntryHash = await alice.cells[0].callZome<EntryHash>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "create",
     payload: content,
   });
   await pause(100);
   const readContent = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
@@ -139,7 +139,7 @@ test("TryCP Scenario - Conductor maintains data after shutdown and restart", asy
   await bob.conductor.startUp({});
   await bob.conductor.connectAppInterface();
   const readContentAfterRestart = await bob.cells[0].callZome<typeof content>({
-    zome_name: "crud",
+    zome_name: "coordinator",
     fn_name: "read",
     payload: createEntryHash,
   });
